WIP split realm-endpoints-test into smaller files
