+++ packages/realm-server/tests/index.ts
@@ -9,7 +9,17 @@ import './loader-test';
 import './module-syntax-test';
 import './permissions/permission-checker-test';
 import './queue-test';
+import './card-endpoints-test';
+import './card-source-endpoints-test';
 import './realm-endpoints-test';
+import './realm-endpoints/directory-test';
+import './realm-endpoints/info-test';
+import './realm-endpoints/mtimes-test';
+import './realm-endpoints/permissions-test';
+import './realm-endpoints/search-test';
+import './realm-endpoints/user-test';
+import './search-prerendered-test';
 import './server-endpoints-test';
+import './types-endpoint-test';
 import './virtual-network-test';
 import './billing-test';
