+++ packages/realm-server/tests/realm-endpoints/directory-test.ts
@@ -1,0 +1,217 @@
+import { module, test } from 'qunit';
+import supertest, { Test, SuperTest } from 'supertest';
+import { join, resolve, basename } from 'path';
+import { Server } from 'http';
+import { dirSync, setGracefulCleanup, type DirResult } from 'tmp';
+import { copySync, ensureDirSync } from 'fs-extra';
+import {
+  baseRealm,
+  Realm,
+  RealmPermissions,
+  type LooseSingleCardDocument,
+} from '@cardstack/runtime-common';
+import {
+  setupCardLogs,
+  setupBaseRealmServer,
+  runTestRealmServer,
+  setupDB,
+  createVirtualNetwork,
+  createVirtualNetworkAndLoader,
+  matrixURL,
+  closeServer,
+} from '../helpers';
+import '@cardstack/runtime-common/helpers/code-equality-assertion';
+import { resetCatalogRealms } from '../../handlers/handle-fetch-catalog-realms';
+
+setGracefulCleanup();
+const testRealmURL = new URL('http://127.0.0.1:4444/');
+const testRealmHref = testRealmURL.href;
+const distDir = resolve(join(__dirname, '..', '..', '..', 'host', 'dist'));
+console.log(`using host dist dir: ${distDir}`);
+
+let createJWT = (
+  realm: Realm,
+  user: string,
+  permissions: RealmPermissions['user'] = [],
+) => {
+  return realm.createJWT(
+    {
+      user,
+      realm: realm.url,
+      permissions,
+      sessionRoom: `test-session-room-for-${user}`,
+    },
+    '7d',
+  );
+};
+
+module(basename(__filename), function () {
+  module('Realm-specific Endpoints | GET directory path', function (hooks) {
+    let testRealm: Realm;
+    let testRealmHttpServer: Server;
+    let request: SuperTest<Test>;
+    let dir: DirResult;
+
+    function setupPermissionedRealm(
+      hooks: NestedHooks,
+      permissions: RealmPermissions,
+      fileSystem?: Record<string, string | LooseSingleCardDocument>,
+    ) {
+      setupDB(hooks, {
+        beforeEach: async (_dbAdapter, publisher, runner) => {
+          dir = dirSync();
+          let testRealmDir = join(dir.name, '..', 'realm_server_1', 'test');
+          ensureDirSync(testRealmDir);
+          // If a fileSystem is provided, use it to populate the test realm, otherwise copy the default cards
+          if (!fileSystem) {
+            copySync(join(__dirname, '..', 'cards'), testRealmDir);
+          }
+
+          let virtualNetwork = createVirtualNetwork();
+
+          ({ testRealm, testRealmHttpServer } = await runTestRealmServer({
+            virtualNetwork,
+            testRealmDir,
+            realmsRootPath: join(dir.name, '..', 'realm_server_1'),
+            realmURL: testRealmURL,
+            permissions,
+            dbAdapter: _dbAdapter,
+            runner,
+            publisher,
+            matrixURL,
+            fileSystem,
+          }));
+
+          request = supertest(testRealmHttpServer);
+        },
+      });
+    }
+
+    let { virtualNetwork, loader } = createVirtualNetworkAndLoader();
+
+    setupCardLogs(
+      hooks,
+      async () => await loader.import(`${baseRealm.url}card-api`),
+    );
+
+    setupBaseRealmServer(hooks, virtualNetwork, matrixURL);
+
+    hooks.beforeEach(async function () {
+      dir = dirSync();
+      copySync(join(__dirname, '..', 'cards'), dir.name);
+    });
+
+    hooks.afterEach(async function () {
+      await closeServer(testRealmHttpServer);
+      resetCatalogRealms();
+    });
+
+    module('public readable realm', function (hooks) {
+      setupPermissionedRealm(hooks, {
+        '*': ['read'],
+      });
+
+      test('serves the request', async function (assert) {
+        let response = await request
+          .get('/dir/')
+          .set('Accept', 'application/vnd.api+json');
+
+        assert.strictEqual(response.status, 200, 'HTTP 200 status');
+        assert.strictEqual(
+          response.get('X-boxel-realm-url'),
+          testRealmURL.href,
+          'realm url header is correct',
+        );
+        assert.strictEqual(
+          response.get('X-boxel-realm-public-readable'),
+          'true',
+          'realm is public readable',
+        );
+        let json = response.body;
+        for (let relationship of Object.values(json.data.relationships)) {
+          delete (relationship as any).meta.lastModified;
+        }
+        assert.deepEqual(
+          json,
+          {
+            data: {
+              id: `${testRealmHref}dir/`,
+              type: 'directory',
+              relationships: {
+                'bar.txt': {
+                  links: {
+                    related: `${testRealmHref}dir/bar.txt`,
+                  },
+                  meta: {
+                    kind: 'file',
+                  },
+                },
+                'foo.txt': {
+                  links: {
+                    related: `${testRealmHref}dir/foo.txt`,
+                  },
+                  meta: {
+                    kind: 'file',
+                  },
+                },
+                'subdir/': {
+                  links: {
+                    related: `${testRealmHref}dir/subdir/`,
+                  },
+                  meta: {
+                    kind: 'directory',
+                  },
+                },
+              },
+            },
+          },
+          'the directory response is correct',
+        );
+      });
+    });
+
+    module('permissioned realm', function (hooks) {
+      setupPermissionedRealm(hooks, {
+        john: ['read'],
+      });
+
+      test('401 with invalid JWT', async function (assert) {
+        let response = await request
+          .get('/dir/')
+          .set('Accept', 'application/vnd.api+json')
+          .set('Authorization', `Bearer invalid-token`);
+
+        assert.strictEqual(response.status, 401, 'HTTP 401 status');
+      });
+
+      test('401 without a JWT', async function (assert) {
+        let response = await request
+          .get('/dir/')
+          .set('Accept', 'application/vnd.api+json'); // no Authorization header
+
+        assert.strictEqual(response.status, 401, 'HTTP 401 status');
+      });
+
+      test('403 without permission', async function (assert) {
+        let response = await request
+          .get('/dir/')
+          .set('Accept', 'application/vnd.api+json')
+          .set('Authorization', `Bearer ${createJWT(testRealm, 'not-john')}`);
+
+        assert.strictEqual(response.status, 403, 'HTTP 403 status');
+      });
+
+      test('200 with permission', async function (assert) {
+        let response = await request
+          .get('/dir/')
+          .set('Accept', 'application/vnd.api+json')
+          .set(
+            'Authorization',
+            `Bearer ${createJWT(testRealm, 'john', ['read'])}`,
+          );
+
+        assert.strictEqual(response.status, 200, 'HTTP 200 status');
+      });
+    });
+  });
+});
