+++ packages/realm-server/tests/realm-endpoints/mtimes-test.ts
@@ -1,0 +1,155 @@
+import { module, test } from 'qunit';
+import supertest, { Test, SuperTest } from 'supertest';
+import { join, resolve, basename } from 'path';
+import { Server } from 'http';
+import { dirSync, setGracefulCleanup, type DirResult } from 'tmp';
+import { copySync, ensureDirSync } from 'fs-extra';
+import {
+  baseRealm,
+  Realm,
+  RealmPermissions,
+  type LooseSingleCardDocument,
+} from '@cardstack/runtime-common';
+import {
+  setupCardLogs,
+  setupBaseRealmServer,
+  runTestRealmServer,
+  setupDB,
+  createVirtualNetwork,
+  createVirtualNetworkAndLoader,
+  matrixURL,
+  closeServer,
+  mtimes,
+} from '../helpers';
+import '@cardstack/runtime-common/helpers/code-equality-assertion';
+
+setGracefulCleanup();
+const testRealmURL = new URL('http://127.0.0.1:4444/');
+const testRealmHref = testRealmURL.href;
+const distDir = resolve(join(__dirname, '..', '..', '..', 'host', 'dist'));
+console.log(`using host dist dir: ${distDir}`);
+
+let createJWT = (
+  realm: Realm,
+  user: string,
+  permissions: RealmPermissions['user'] = [],
+) => {
+  return realm.createJWT(
+    {
+      user,
+      realm: realm.url,
+      permissions,
+      sessionRoom: `test-session-room-for-${user}`,
+    },
+    '7d',
+  );
+};
+
+module(basename(__filename), function () {
+  module('Realm-specific Endpoints | GET _mtimes', function (hooks) {
+    let testRealm: Realm;
+    let testRealmPath: string;
+    let testRealmHttpServer: Server;
+    let request: SuperTest<Test>;
+    let dir: DirResult;
+
+    function setupPermissionedRealm(
+      hooks: NestedHooks,
+      permissions: RealmPermissions,
+      fileSystem?: Record<string, string | LooseSingleCardDocument>,
+    ) {
+      setupDB(hooks, {
+        beforeEach: async (_dbAdapter, publisher, runner) => {
+          dir = dirSync();
+          let testRealmDir = join(dir.name, '..', 'realm_server_1', 'test');
+          ensureDirSync(testRealmDir);
+          // If a fileSystem is provided, use it to populate the test realm, otherwise copy the default cards
+          if (!fileSystem) {
+            copySync(join(__dirname, '..', 'cards'), testRealmDir);
+          }
+
+          let virtualNetwork = createVirtualNetwork();
+
+          ({
+            testRealm,
+            testRealmHttpServer,
+            testRealmDir: testRealmPath,
+          } = await runTestRealmServer({
+            virtualNetwork,
+            testRealmDir,
+            realmsRootPath: join(dir.name, '..', 'realm_server_1'),
+            realmURL: testRealmURL,
+            permissions,
+            dbAdapter: _dbAdapter,
+            runner,
+            publisher,
+            matrixURL,
+            fileSystem,
+          }));
+
+          request = supertest(testRealmHttpServer);
+        },
+      });
+    }
+
+    let { virtualNetwork, loader } = createVirtualNetworkAndLoader();
+
+    setupCardLogs(
+      hooks,
+      async () => await loader.import(`${baseRealm.url}card-api`),
+    );
+
+    setupBaseRealmServer(hooks, virtualNetwork, matrixURL);
+
+    setupPermissionedRealm(hooks, {
+      mary: ['read'],
+    });
+
+    hooks.beforeEach(async function () {
+      dir = dirSync();
+      copySync(join(__dirname, '..', 'cards'), dir.name);
+    });
+
+    hooks.afterEach(async function () {
+      await closeServer(testRealmHttpServer);
+    });
+
+    test('non read permission GET /_mtimes', async function (assert) {
+      let response = await request
+        .get('/_mtimes')
+        .set('Accept', 'application/vnd.api+json')
+        .set('Authorization', `Bearer ${createJWT(testRealm, 'not-mary')}`);
+
+      assert.strictEqual(response.status, 403, 'HTTP 403 status');
+    });
+
+    test('read permission GET /_mtimes', async function (assert) {
+      let expectedMtimes = mtimes(testRealmPath, testRealmURL);
+      delete expectedMtimes[`${testRealmURL}.realm.json`];
+
+      let response = await request
+        .get('/_mtimes')
+        .set('Accept', 'application/vnd.api+json')
+        .set(
+          'Authorization',
+          `Bearer ${createJWT(testRealm, 'mary', ['read'])}`,
+        );
+
+      assert.strictEqual(response.status, 200, 'HTTP 200 status');
+      let json = response.body;
+      assert.deepEqual(
+        json,
+        {
+          data: {
+            type: 'mtimes',
+            id: testRealmHref,
+            attributes: {
+              mtimes: expectedMtimes,
+            },
+          },
+        },
+        'mtimes response is correct',
+      );
+    });
+  });
+});
